% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/Install_And_Load.R
\name{Install_And_Load}
\alias{Install_And_Load}
\title{Install_And_Load Function}
\usage{
Install_And_Load(Required_Packages)
}
\arguments{
\item{Required_Packages}{}
}
\description{
This function allows you to install r packages in an easy way
}
\examples{
# adminpackage4r
# Specify the list of required packages to be installed and load
Required_Packages=c("package1", "package2", "package_n");
# Call the Function
Install_And_Load(Required_Packages);
}
\keyword{Required_Packages}

